Move vue form to generic, deal with defaults
